- moves async documentation tests to function
